Corrected bug on *ballistic demos. I needed to update startTime to double

+++ src/demos/bigballistic/bigballistic.cpp
@@ -30,7 +30,7 @@ class AmmoRound : public cyclone::CollisionSphere
 {
 public:
     ShotType type;
-    unsigned startTime;
+    double startTime;
 
     AmmoRound()
     {
